Moved guts to pkg
