+++ pkg/guts/cost_function.go
