+++ pkg/guts/delta.go
