+++ pkg/guts/deltas.go
