+++ pkg/guts/feed_forward.go
