+++ pkg/guts/layer.go
