+++ pkg/guts/neuron.go
