+++ pkg/guts/sigmoid.go
